+++ app/services/systemlogs/SystemLogEventsService.scala
@@ -4,7 +4,10 @@ import domain.systemlogs.SystemLogEvents
 import services.CrudService
 import services.systemlogs.Impl.SystemLogEventsServiceImpl
 
+import scala.concurrent.Future
+
 trait SystemLogEventsService extends CrudService[SystemLogEvents]{
+  def getSiteLogEvents(siteId: String): Future[Seq[SystemLogEvents]]
 
 }
